Split NFT component to subcomponents
